Various improvements (#290)
* Split admm module into submodules

* Change submodule structure

* Split pgm module into submodules

* Style fix

* Trivial changes

* Typo fix

* Style fixes

* Add missing typing annotations and resolve resulting error

* Docstring fixes
